+++ packages/workbox-range-requests/src/Plugin.ts
@@ -6,9 +6,9 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {createPartialResponse} from './createPartialResponse.mjs';
+import {createPartialResponse} from './createPartialResponse.js';
 
-import './_version.mjs';
+import './_version.js';
 
 /**
  * The range request plugin makes it easy for a request with a 'Range' header to
